refactor: 💡 more a11y
