+++ apps/cf-pages/src/components/molecules/emptyroom/EmptyRoom.stories.tsx
