+++ apps/cf-pages/src/components/templates/Top/Top.tsx
@@ -1,5 +1,5 @@
 import { pageTitle } from "@/components/meta";
-import { EmptyRoom } from "@/components/organisms/emptyroom";
+import { EmptyRoom } from "@/components/molecules/emptyroom";
 import { Main } from "@/components/organisms/main";
 import Head from "next/head";
 import styles from "./styles.module.css";
